+++ src/include/fst/script/arciterator-class.h
@@ -24,12 +24,12 @@ namespace script {
 class ArcIteratorImplBase {
  public:
   virtual bool Done() const = 0;
-  virtual uint32 Flags() const = 0;
+  virtual uint8 Flags() const = 0;
   virtual void Next() = 0;
   virtual size_t Position() const = 0;
   virtual void Reset() = 0;
   virtual void Seek(size_t a) = 0;
-  virtual void SetFlags(uint32 flags, uint32 mask) = 0;
+  virtual void SetFlags(uint8 flags, uint8 mask) = 0;
   virtual ArcClass Value() const = 0;
   virtual ~ArcIteratorImplBase() {}
 };
@@ -43,7 +43,7 @@ class ArcIteratorClassImpl : public ArcIteratorImplBase {
 
   bool Done() const final { return aiter_.Done(); }
 
-  uint32 Flags() const final { return aiter_.Flags(); }
+  uint8 Flags() const final { return aiter_.Flags(); }
 
   void Next() final { aiter_.Next(); }
 
@@ -53,9 +53,7 @@ class ArcIteratorClassImpl : public ArcIteratorImplBase {
 
   void Seek(size_t a) final { aiter_.Seek(a); }
 
-  void SetFlags(uint32 flags, uint32 mask) final {
-    aiter_.SetFlags(flags, mask);
-  }
+  void SetFlags(uint8 flags, uint8 mask) final { aiter_.SetFlags(flags, mask); }
 
   // This is returned by value because it has not yet been constructed, and
   // is likely to participate in return-value optimization.
@@ -83,7 +81,7 @@ class ArcIteratorClass {
 
   bool Done() const { return impl_->Done(); }
 
-  uint32 Flags() const { return impl_->Flags(); }
+  uint8 Flags() const { return impl_->Flags(); }
 
   void Next() { impl_->Next(); }
 
@@ -93,7 +91,7 @@ class ArcIteratorClass {
 
   void Seek(size_t a) { impl_->Seek(a); }
 
-  void SetFlags(uint32 flags, uint32 mask) { impl_->SetFlags(flags, mask); }
+  void SetFlags(uint8 flags, uint8 mask) { impl_->SetFlags(flags, mask); }
 
   ArcClass Value() const { return impl_->Value(); }
 
@@ -131,7 +129,7 @@ class MutableArcIteratorClassImpl
 
   bool Done() const final { return aiter_.Done(); }
 
-  uint32 Flags() const final { return aiter_.Flags(); }
+  uint8 Flags() const final { return aiter_.Flags(); }
 
   void Next() final { aiter_.Next(); }
 
@@ -141,9 +139,7 @@ class MutableArcIteratorClassImpl
 
   void Seek(size_t a) final { aiter_.Seek(a); }
 
-  void SetFlags(uint32 flags, uint32 mask) final {
-    aiter_.SetFlags(flags, mask);
-  }
+  void SetFlags(uint8 flags, uint8 mask) final { aiter_.SetFlags(flags, mask); }
 
   void SetValue(const Arc &arc) { aiter_.SetValue(arc); }
 
@@ -175,7 +171,7 @@ class MutableArcIteratorClass {
 
   bool Done() const { return impl_->Done(); }
 
-  uint32 Flags() const { return impl_->Flags(); }
+  uint8 Flags() const { return impl_->Flags(); }
 
   void Next() { impl_->Next(); }
 
@@ -185,7 +181,7 @@ class MutableArcIteratorClass {
 
   void Seek(size_t a) { impl_->Seek(a); }
 
-  void SetFlags(uint32 flags, uint32 mask) { impl_->SetFlags(flags, mask); }
+  void SetFlags(uint8 flags, uint8 mask) { impl_->SetFlags(flags, mask); }
 
   void SetValue(const ArcClass &ac) { impl_->SetValue(ac); }
